[review] new page (기록남기기) props 정리
